Fixed the previous issue : the script was cycling through the first and second page. On the first page there's only a "next" button to go to the second page, however there was a "previous" button on the second page and the script followed it. Now the pages already crawled are stored inside the "crawled_pages" list, and the script doesn't crawl through links that are already in this list.
You can also choose the number of items to scrape in the CLOSESPIDER_ITEMCOUNT parameter in settings.py. Setting this parameter to 0 means the number of items to scrape is not limited.

New issue : repetitions in the lists, crawling of the very last page while we want to crawl them in growing order.